+++ test/test_keeper_request_wire_view.ml
@@ -64,7 +64,7 @@ let view ~front history : Try_provider.request_view =
     ~last_resort:false
     ~base_path:""
     ~demote_before:0
-    ~completed_end_atom:0
+    ~turn_boundary:(Masc.Keeper_carried_front.Turn_boundary { end_atom = 0 })
     ~materialize:(fun ~pending:_ messages -> messages)
     history
 ;;